agregando metodo get y delete
