+++ src/main/java/com/example/stose/controller/LibroController.java
@@ -31,13 +31,6 @@ public class LibroController {
 
 		return "libros";
 	}
-	
-	@GetMapping({ "/libros/info/{id}"})
-	public String infoLibros(@PathVariable Long id, Model modelo) {
-		modelo.addAttribute("libro", servicio.obtenerLibroPorId(id));
-
-		return "detalle-libro";
-	}
 
 	@GetMapping("/libros/nuevo")
 	public String mostrarFormularioDeRegistrarLibro(Model modelo) {
@@ -81,31 +74,6 @@ public class LibroController {
 		servicio.actualizarLibro(libroExistente);
 		return "redirect:/inicio/libros";
 	}
-	
-	@PostMapping("/libros/detalle/{id}")
-	public String detalleLibro(@PathVariable Long id, @ModelAttribute("libro") Libro libro,
-			Model modelo) {
-		Libro libroExistente = servicio.obtenerLibroPorId(id);
-		libroExistente.setId(id);
-		libroExistente.setImg(libro.getImg());
-		libroExistente.setTitulo(libro.getTitulo());
-		libroExistente.setAutor(libro.getAutor());
-		libroExistente.setId_editorial(libro.getId_editorial());
-		libroExistente.setId_categoria(libro.getId_categoria());
-		libroExistente.setId_subcategoria(libro.getId_subcategoria());
-		libroExistente.setDescripcion(libro.getDescripcion());
-		libroExistente.setPrecio_original(libro.getPrecio_original());
-		libroExistente.setPrecio_descuento(libro.getPrecio_descuento());
-		libroExistente.setPorcentaje_descuento(libro.getPorcentaje_descuento());
-		libroExistente.setNumero_pagina(libro.getNumero_pagina());
-		libroExistente.setYear_publicacion(libro.getYear_publicacion());
-		libroExistente.setDisponible(libro.getDisponible());
-		libroExistente.setDisponible(libro.getDescripcion());
-
-		servicio.actualizarLibro(libroExistente);
-		return "redirect:/inicio/tienda";
-	}
-
 
 	@GetMapping("/libros/{id}")
 	public String eliminarLibro(@PathVariable Long id) {
