+++ src/main/resources/templates/costos-de-envio.html
@@ -179,7 +179,7 @@
                                                                 <span class="fly-item">x<span th:text="${carrito.cantidad}"></span></span>
                                                             </span>
                                                         </div>
-                                                        <a th:href="@{/inicio/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
+                                                        <a th:href="@{/inicio/costos-de-envio/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
                                                     </li>
                                                 </ul>
                                             </div>
