+++ src/main/resources/templates/libro-oferta.html
@@ -181,7 +181,7 @@
                                                             <span class="fly-item">x<span th:text="${carrito.cantidad}"></span></span>
                                                         </span>
                                                     </div>
-                                                    <a th:href="@{/inicio/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
+                                                    <a th:href="@{/inicio/libro-oferta/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
                                                 </li>
                                             </ul>
                                         </div>
