+++ src/main/resources/templates/lista-de-deseos.html
@@ -183,7 +183,7 @@
                                                             <span class="fly-item">x<span th:text="${carrito.cantidad}"></span></span>
                                                         </span>
                                                     </div>
-                                                    <a th:href="@{/inicio/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
+                                                    <a th:href="@{/inicio/lista-de-deseos/{id}(id=${carrito.id})}" class="item-remove"><i class="ri-close-line"></i></a>
                                                 </li>
                                             </ul>
                                         </div>
